% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cohortOverview.R
\name{cohortOverview_tab}
\alias{cohortOverview_tab}
\title{Cohort overview tab UI}
\usage{
cohortOverview_tab(
  sample_selection,
  gene_selection,
  profile_variables,
  colour_variables,
  id = NULL
)
}
\arguments{
\item{sample_selection}{radio input(s) for sample class(es)}

\item{gene_selection}{select input with gene symbols}

\item{profile_variables}{list of group variables for trajectory profile}

\item{colour_variables}{list of unique variables for trajectory color}

\item{id}{optional module ID}
}
\value{
tab panel with inputs
}
\description{
Cohort overview tab UI
}
\examples{
if (interactive()) { 
sample_selection <- radioButtons("sample_1", "Select sample", samples)
gene_selection <- selectizeInput("gene_selection", "Select gene", gene_list)
cohortOverview_tab(sample_selection, gene_selection,
 c("measureA", "measureB"),
 c("measureA_1", "measureA_2"))
}
}
