% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_compute.R
\name{correlationResultsToLong}
\alias{correlationResultsToLong}
\title{Transform correlation data frame to long format}
\usage{
correlationResultsToLong(
  data,
  first_col_name = "Gene",
  name_to = "ClinicalVariable"
)
}
\arguments{
\item{data}{correlation data frame from \code{\link[=correlateMatrices]{correlateMatrices()}}}

\item{first_col_name}{name to set the first column (colnames of matrix)}

\item{name_to}{name of column for long format}
}
\value{
long format data frame
}
\description{
Transform correlation data frame to long format
}
