% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_compute.R
\name{corrResultsToTable}
\alias{corrResultsToTable}
\title{Collapse data frame into table with highlights}
\usage{
corrResultsToTable(df, max_pvalue = 0.05, use_padj = F, rowname_col = "Gene")
}
\arguments{
\item{df}{wide correlation data frame}

\item{max_pvalue}{p-value filter for highlight}

\item{use_padj}{flag for normal or adjusted p-value}

\item{rowname_col}{column that contains rownames}
}
\value{
data frame with HTML b tags
}
\description{
Collapse data frame into table with highlights
}
