% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_config.R
\name{create_config_template}
\alias{create_config_template}
\title{Create a barebones configuration file template}
\usage{
create_config_template(filename = "config.yaml", target_dir = NULL)
}
\arguments{
\item{filename}{optional filename, default is config.yaml}

\item{target_dir}{optional target directory, default is current working
directory}
}
\value{
Creates configuration file in \code{target_dir}
}
\description{
The resulting file contain placeholder text in all caps for you to edit
according to your needs. It also includes the three correlation modules.
}
\examples{
if (interactive()) {
create_config_template()
}

}
