% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_example.R
\name{create_example}
\alias{create_example}
\title{Create example files}
\usage{
create_example(target_dir = NULL)
}
\arguments{
\item{target_dir}{(optional) a directory where to create the files, otherwise uses
the output of getwd()}
}
\value{
Create examples files in target_dir
}
\description{
Create example files
}
\examples{
if (interactive()) {
dir.create("newapp")
create_example("newapp")
}
}
