% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mod_template.R
\name{create_module_template}
\alias{create_module_template}
\title{Creates a module code template in current working directory}
\usage{
create_module_template(module_name)
}
\arguments{
\item{module_name}{module name in camelCase}
}
\description{
Creates a module code template in current working directory
}
