% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_degVolcano.R
\name{degVolcano_tab}
\alias{degVolcano_tab}
\title{Differentially expressed genes tab UI}
\usage{
degVolcano_tab(model_names, id = NULL)
}
\arguments{
\item{model_names}{list of model names}

\item{id}{optional module ID}
}
\value{
tab panel with inputs
}
\description{
Differentially expressed genes tab UI
}
