% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_compute.R
\name{fast_cor}
\alias{fast_cor}
\title{Parallel computation of correlation of data frame vs matrix}
\usage{
fast_cor(x, y, method = c("pearson", "spearman", "kendall"), cores = 1)
}
\arguments{
\item{x}{numeric data frame}

\item{y}{matrix}

\item{method}{correlation method to use}

\item{cores}{number of cores to use with the \code{\link[parallel:mclapply]{parallel::mclapply()}} function}
}
\value{
numeric matrix
}
\description{
Uses mclapply so it won't run in parallel on Windows
}
