% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_geneModulesCorr.R
\name{geneModulesCorr_tab}
\alias{geneModulesCorr_tab}
\title{Gene modules tab UI}
\usage{
geneModulesCorr_tab(
  sample_select,
  sources_names,
  clinical_variables,
  id = NULL
)
}
\arguments{
\item{sample_select}{radio inputs for sample classes}

\item{sources_names}{list of module sources}

\item{clinical_variables}{list of set of clinical variables for scatterplots}

\item{id}{optional module ID}
}
\value{
tab panel with inputs
}
\description{
Gene modules tab UI
}
