% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_compute.R
\name{longCorrelationMatrix}
\alias{longCorrelationMatrix}
\title{Compute correlation data frame in long format}
\usage{
longCorrelationMatrix(
  first_col_name = "Gene",
  name_to = "ClinicalVariable",
  ...
)
}
\arguments{
\item{first_col_name}{name to set the first column (colnames of matrix)}

\item{name_to}{name of column for long format}

\item{...}{arguments to \code{\link[=correlateMatrices]{correlateMatrices()}}}
}
\value{
long format data frame
}
\description{
Compute correlation data frame in long format
}
