% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(config_file, data_folder = "", ...)
}
\arguments{
\item{config_file}{The name of the yaml configuration file}

\item{data_folder}{Optional directory prefix for data files.
Enables using the same configuration file for different versions of data files.}

\item{...}{Further optional arguments.}
}
\description{
Run the Shiny Application
}
