% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{show_available_modules}
\alias{show_available_modules}
\title{Print list of currently supported modules}
\usage{
show_available_modules()
}
\value{
list of available modules
}
\description{
See \code{vignette("config", package = "shinyExprPortal")} for details on how
to configure each module.
}
\examples{
show_available_modules()
}
