% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_singleGeneCorr.R
\name{singleGeneCorr_tab}
\alias{singleGeneCorr_tab}
\title{Single gene correlation tab UI}
\usage{
singleGeneCorr_tab(
  sample_select,
  gene_select,
  colours,
  outputs,
  advanced = TRUE,
  id = NULL
)
}
\arguments{
\item{sample_select}{radio inputs for sample classes}

\item{gene_select}{select input with gene symbols}

\item{colours}{list of variables for colour selection}

\item{outputs}{configuration of tabs with plots}

\item{advanced}{boolean flag to show or hide advanced options
such as outlier removal}

\item{id}{optional module ID}
}
\value{
tab panel with inputs
}
\description{
Single gene correlation tab UI
}
\examples{
if (interactive()) {
 sample_select <- radioButtons("sample_1", "Select sample", samples)
 gene_select <- selectizeInput("gene_selection", "Select gene", gene_list)
 output_list <- list("Vars" = c("Var1", "Var2"))
 singleGeneCorr_tab(sample_select, gene_select, c("Var1", "Var2"),
  output_list)
}
}
