% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_singleVariableCorr.R
\name{singleVariableCorr_tab}
\alias{singleVariableCorr_tab}
\title{All genes correlation tab UI}
\usage{
singleVariableCorr_tab(
  sample_select,
  clinical_variables,
  advanced = TRUE,
  id = NULL
)
}
\arguments{
\item{sample_select}{radio inputs for sample classes}

\item{clinical_variables}{select input for clinical variables}

\item{advanced}{boolean flag to show or hide advanced options}

\item{id}{optional module ID}
}
\value{
tab panel with inputs
}
\description{
All genes correlation tab UI
}
