% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_vegaplots.R
\name{vegalite_volcanoplot}
\alias{vegalite_volcanoplot}
\title{Vega-lite volcanoplot}
\usage{
vegalite_volcanoplot(
  data,
  fc_min = 1,
  pvalue_min = 0.05,
  pvalue_col = "P.value",
  gene_col = "Gene",
  colors = c("black", "green", "blue", "red"),
  strokeDash = c(4, 4),
  opacity = 0.5
)
}
\arguments{
\item{data}{DE table enriched with color column}

\item{fc_min}{fold change significance line. Default is 1 (and -1)}

\item{pvalue_min}{p-value significance line. Default is 0.05}

\item{pvalue_col}{column to use for p-value position. Default is P.value}

\item{colors}{optional list of custom colors for not significant, logFC,
p-value and logFC and p-value significant. Defaults are black, blue, green
and red}

\item{strokeDash}{vector with strokeDash configuration. Default is 4,4}

\item{opacity}{float value for points opacity. Default is 0.5}
}
\value{
vega-lite R spec
}
\description{
The table must be enriched with a column called color that contains
}
