% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_wholeDataCorr.R
\name{wholeDataCorr_tab}
\alias{wholeDataCorr_tab}
\title{Title}
\usage{
wholeDataCorr_tab(
  sample_select,
  clinical_variables,
  advanced = NULL,
  id = NULL
)
}
\arguments{
\item{sample_select}{a}

\item{clinical_variables}{b}

\item{advanced}{advanced options?}

\item{id}{optional module ID}
}
\value{
a tab panel
}
\description{
Title
}
