% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_config.R
\name{create_config_template}
\alias{create_config_template}
\title{Create a bare-bones configuration file template}
\usage{
create_config_template(target_dir, filename = "config.yaml")
}
\arguments{
\item{target_dir}{location to create the configuration file}

\item{filename}{optional file name, default is config.yaml}
}
\value{
Creates configuration file in \code{target_dir}
}
\description{
The resulting file contain placeholder text in upper case for you to edit
according to your needs. It also includes the three correlation modules by
default.
}
\examples{
if (interactive()) {
    dir.create("newapp")
    create_config_template("newapp")
}

}
