% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_example.R
\name{create_example}
\alias{create_example}
\title{Create example files}
\usage{
create_example(target_dir)
}
\arguments{
\item{target_dir}{location where to create the files}
}
\value{
Create examples files in target_dir
}
\description{
Create example files for measures, expression matrix and lookup table
}
\examples{
if (interactive()) {
    dir.create("newapp")
    create_example("newapp")
}
}
