% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_degOverview.R
\name{degOverview_tab}
\alias{degOverview_tab}
\title{Differentially expressed genes tab UI}
\usage{
degOverview_tab(categories, id = NULL)
}
\arguments{
\item{categories}{model categories}

\item{id}{optional module ID}
}
\value{
tab panel with inputs
}
\description{
Differentially expressed genes tab UI
}
